FINAL COMMIT CARALHO

+++ disk-usage/src/arguments.h
@@ -22,7 +22,7 @@ struct Arguments {
     bool all;
     bool bytes;
     bool block_size_flag;
-    int block_size;
+    off_t block_size;
     bool countLinks;
     bool deference;
     bool separateDirs;
